requirements.txt
# Core dependencies
numpy>=1.24.0
pandas>=2.0.0
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# Image processing
Pillow>=10.0.0
opencv-python>=4.8.0

# Audio processing
librosa>=0.10.0
soundfile>=0.12.0

# Retrieval and vector databases
faiss-cpu>=1.7.4
chromadb>=0.4.0

# API and web
flask>=2.3.0
flask-cors>=4.0.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
scikit-learn>=1.3.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

